Attempt at dry-er form was a very bad idea.
