+++ src/index.jsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import App from "./components/App";
 import { Provider } from "react-redux";
 import * as serviceWorker from "./serviceWorker";
-import store from "./store";
+import store from "./store.js";
 import "./index.css";
 import "bootstrap/dist/css/bootstrap.min.css";
